+++ firebase/firebase.json
@@ -20,5 +20,11 @@
     "hosting": {
       "port": "5000"
     }
+  },
+  "functions": {
+    "predeploy": [
+      "npm --prefix \"$RESOURCE_DIR\" run lint",
+      "npm --prefix \"$RESOURCE_DIR\" run build"
+    ]
   }
 }
