+++ firebase/functions/credentials/stripe_credentials.ts
@@ -1,0 +1,4 @@
+export const api_keys = {
+    "dev_api_key": "sk_test_7zn5A84TajuMWEBE8brtaDPi00MHzyzMr1",
+    "live_api_key": "sk_live_LkcABzpcdU42EDt5GK5ze6X200pGAgBhdw"
+}
