+++ firebase/functions/package.json
@@ -2,26 +2,34 @@
   "name": "functions",
   "description": "Cloud Functions for Firebase",
   "scripts": {
-    "serve": "firebase serve --only functions",
+    "build": "./node_modules/.bin/tslint --project tsconfig.json && ./node_modules/.bin/tsc",
+    "serve": "npm run build && firebase serve --only functions",
     "shell": "firebase functions:shell",
     "start": "npm run shell",
     "deploy": "firebase deploy --only functions",
     "logs": "firebase functions:log",
+    "lint": "tslint -p tsconfig.json src/**/*.ts -e tests/**/*.ts",
     "test": "mocha --reporter spec"
   },
   "engines": {
     "node": "8"
   },
   "dependencies": {
+    "@types/express": "^4.17.6",
+    "@types/firebase": "^3.2.1",
+    "@types/stripe": "^7.13.23",
     "acuityscheduling": "^0.1.9",
     "firebase-admin": "^8.6.1",
     "firebase-functions": "^3.0.0",
     "googleapis": "^41.0.1",
-    "stripe": "^8.62.0"
+    "stripe": "^8.62.0",
+    "tslint": "^6.1.2",
+    "typescript": "^3.9.5"
   },
   "devDependencies": {
     "firebase-functions-test": "^0.1.7",
     "mocha": "^6.2.2"
   },
+  "main": "lib/src/index.js",
   "private": true
 }
