+++ firebase/functions/src/environment.d.ts
@@ -1,0 +1,9 @@
+declare global {
+  namespace NodeJS {
+    interface ProcessEnv {
+      FIREBASE_CONFIG: string;
+    }
+  }
+}
+
+export {}
