+++ firebase/package.json
@@ -7,16 +7,16 @@
     "@material-ui/core": "^4.9.11",
     "@material-ui/icons": "^4.9.1",
     "@material-ui/pickers": "^3.2.10",
-    "date-fns": "^2.0.0-beta.5",
+    "autoprefixer": "9.8.0",
+    "date-fns": "^2.14.0",
     "dateformat": "^3.0.3",
-    "firebase": "^6.6.2",
+    "firebase": "^7.15.3",
     "googleapis": "^39.2.0",
     "moment": "^2.26.0",
     "moment-timezone": "^0.5.28",
     "query-string": "^6.8.3",
     "react": "^16.10.2",
     "react-dom": "^16.10.2",
-    "react-final-form": "^6.3.0",
     "react-loading-overlay": "^1.0.1",
     "react-router-dom": "^5.1.2",
     "react-scripts": "^3.4.1",
